requirements.txt
# Smart Ad Planner Dependencies

# Web Framework
fastapi==0.104.1
sse-starlette==1.8.2
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# AI/ML - Google Vertex AI (supports ADC)
google-cloud-aiplatform>=1.127.0

# Authentication - Application Default Credentials
google-auth>=2.23.4
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Vector Database
chromadb==0.4.18

# PDF Generation
reportlab==4.0.7

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Dashboard & Visualization
streamlit==1.29.0

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-playwright==0.4.3
playwright==1.40.0
